blog2doc/requirements.txt
lxml=3.7.3
requests=2.13.0
python-docx=0.8.6
